Updated code to fix removing resumable object from array on upload completion

+++ src/app/file-upload-home/file-upload-home.component.ts
@@ -101,12 +101,11 @@ export class FileUploadHomeComponent implements OnInit {
             if(transaction.uid === resumableObject.transactionUid) {
               resumableObject.on('progress', function() {
                 transaction['uploadProgress'] = Math.round(resumableObject.progress(true) * 100);
-                if(transaction['uploadProgress'] === 100) {
+                if(transaction['uploadProgress'] === 100 && !resumableObject.isUploading()) {
+                  self.initUpstreamUpload(resumableObject);
                   self.fileUploadDataService.removeResumableObject(resumableObject);
                 }
               });
-
-              this.initUpstreamUpload(resumableObject);
             }
           });
         });
